requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
fal-client==0.5.6
supabase==2.11.0
python-multipart==0.0.20
python-dotenv==1.0.1
